+++ AssemblyLib/Remapper/Filters/FieldTypeFilters.cs
@@ -2,7 +2,7 @@
 using AssemblyLib.Models;
 using SPTarkov.DI.Annotations;
 
-namespace AssemblyLib.ReMapper.Filters;
+namespace AssemblyLib.Remapper.Filters;
 
 [Injectable(TypePriority = 2)]
 public sealed class FieldTypeFilters : IRemapFilter
