+++ AssemblyLib/Remapper/Metadata/AttributeFactory.cs
@@ -6,12 +6,12 @@ using AsmResolver.DotNet.Signatures.Types;
 using AsmResolver.PE.DotNet.Cil;
 using AsmResolver.PE.DotNet.Metadata.Tables.Rows;
 using AssemblyLib.Models;
-using AssemblyLib.Utils;
+using AssemblyLib.Shared;
 using Serilog;
 using Serilog.Events;
 using SPTarkov.DI.Annotations;
 
-namespace AssemblyLib.ReMapper.MetaData;
+namespace AssemblyLib.Remapper.MetaData;
 
 [Injectable]
 public class AttributeFactory(DataProvider dataProvider)
