+++ AssemblyTool/Commands/RegenerateSignature.cs
@@ -1,7 +1,4 @@
-﻿using AssemblyLib;
-using AssemblyLib.AutoMatcher;
-using AssemblyLib.Utils;
-using AssemblyTool.Utils;
+﻿using AssemblyLib.Shared;
 using CliFx.Attributes;
 using CliFx.Infrastructure;
 
@@ -36,8 +33,6 @@ public class RegenerateSignature : CliFx.ICommand
 
     public async ValueTask ExecuteAsync(IConsole console)
     {
-        Debugger.TryWaitForDebuggerAttach();
-
         var app = new App();
         await app.RunAutoMatcher(AssemblyPath, OldAssemblyPath!, OldTypeName, NewTypeName, true);
     }
